+++ lib/determinator/control.rb
@@ -58,7 +58,7 @@ module Determinator
     Indicators = Struct.new(:rollout, :variant)
 
     def determinate_and_notice(name, id:, guid:, properties:)
-      feature = retrieval.retrieve(name)
+      feature = Determinator.with_retrieval_cache(name) { retrieval.retrieve(name) }
 
       determinate(feature, id: id, guid: guid, properties: properties).tap do |determination|
         Determinator.notice_determination(id, guid, feature, determination)